// src/app/components/tms/model/driver-info.model.ts
export class DriverInfoCreateModel {
    infoId: number;
    employeeName: string;
    employeeId: number;
    contactNo: string;
    companyId: number;
    companyName: string;
    departmentName: string;
    designation: string;
    position: string;
    workLocation: string;
    grade: string;
    joiningDate: Date;
    employeeStatus: string;
    mobileNo: string;
    status: string;
    weekend: string;
    licenseNo: string;
    licenseExpiryDate: Date;
    daysToRemind: number;
    nId: number;
}